youtube_influencers/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.10.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.11
clang==5.0
click==8.0.3
dataclasses==0.6
deepface==0.0.72
dlib==19.17.0
filelock==3.4.1
Flask==2.0.2
flatbuffers==1.12
gast==0.4.0
gdown==4.2.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
imutils==0.5.2
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.1.2.30
opt-einsum==3.3.0
pandas==1.1.5
pika==1.2.0
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.9
rsa==4.8
six==1.15.0
soupsieve==2.3.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.2
wget==3.2
wrapt==1.12.1
zipp==3.6.0
